chatbot_solution/pyproject.toml
[tool.poetry]
name = "chatbot-solution"
version = "0.1.0"
description = "Chatbot for Abastores"
authors = ["Daniel Espinoza <danielx.1997@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.32.2"
python-dotenv = "^1.0.1"
streamlit = "^1.35.0"
openai = "^1.30.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
